ImageStore: Minio store using worker pool
For every image stored into Minio, a go routine is spawned which writes
to the minio db. If the ingestion is very fast, then the number of go
routines will become unbounded.

This patch adds a worker pool of 100 go workers and uses a channel to
transfer data to worker routines. If all workers are busy, the Store
operation will wait.

This is a potential fix for the long run issue involving ImageStore.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>