ImageStore: Fix for minio retention policy not working
The issue is due to not breaking from the infinite for loop while reading
from io.Reader object whenever error occured here  which was in turn resulting
in the increase of more active of ImageStore's Read gRPC call, this has
been fixed by breaking out of the loop. Also, added in few optimizations
(using ticker) to address the minio retention policy by just retuning and not
exiting from the retention policy `go routine` as it exits from the imagestore
program.

Signed-off-by: Vinod K B <vinod.k.b@intel.com>