ImageStore: Fix memory leak with remote client
This patch fixes the memory leak seen when running visualizer with 6
streams on a different machine.
Issue was happening because multiple subscribers open the server in
streaming mode concurrently. This patch limits the MaxConcurrentStreams
to 1 which fixes the memory leak.

Also removes the manual sleep for minio buffer pool as it is not
required when using buffered channels.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>